requirements.txt
fastapi[all]
mysql-connector
requests
spacy